refactor!: Standardize namespace and add CI for API docs
