+++ integrations/uptrain/pydoc/config.yml
@@ -1,0 +1,30 @@
+loaders:
+  - type: haystack_pydoc_tools.loaders.CustomPythonLoader
+    search_path: [../src]
+    modules:
+      [
+        "haystack_integrations.components.evaluators.uptrain.evaluator",
+        "haystack_integrations.components.evaluators.uptrain.metrics",
+      ]
+    ignore_when_discovered: ["__init__"]
+processors:
+  - type: filter
+    expression:
+    documented_only: true
+    do_not_filter_modules: false
+    skip_empty_modules: true
+  - type: smart
+  - type: crossref
+renderer:
+  type: haystack_pydoc_tools.renderers.ReadmePreviewRenderer
+  excerpt: UpTrain integration for Haystack
+  category_slug: haystack-integrations
+  title: UpTrain
+  slug: integrations-uptrain
+  order: 1
+  markdown:
+    descriptive_class_title: false
+    descriptive_module_title: true
+    add_method_class_prefix: true
+    add_member_class_prefix: false
+    filename: _readme_uptrain.md
