--- integrations/uptrain/src/haystack_integrations/components/evaluators/__init__.py
@@ -1,7 +1,0 @@
-from .uptrain_evaluator import UpTrainEvaluator
-from .uptrain_metrics import UpTrainMetric
-
-__all__ = (
-    "UpTrainEvaluator",
-    "UpTrainMetric",
-)
