+++ integrations/uptrain/src/haystack_integrations/components/evaluators/uptrain/__init__.py
@@ -1,0 +1,7 @@
+from .evaluator import UpTrainEvaluator
+from .metrics import UpTrainMetric
+
+__all__ = (
+    "UpTrainEvaluator",
+    "UpTrainMetric",
+)
